requirements.txt
streamlit==1.38.0
google-generativeai
pillow
numpy<2.0.0
ultralytics
streamlit-drawable-canvas